requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
bcrypt==3.2.2
boto3==1.40.62
botocore==1.40.62
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==41.0.7
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==3.2.3
django-storages==1.14.6
djangorestframework==3.16.1
frozenlist==1.8.0
idna==3.11
jmespath==1.0.1
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycares==4.11.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2024.1
requests==2.31.0
rich==13.7.1
s3transfer==0.14.0
sentry-sdk==1.39.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
